Move EM base to object base
